9535: split formattedTrialSessionDetails.test.js further
